refactor: remove index-types.d.ts
Removes extra file operations by using `export namespace`.
